+++ index.js
@@ -27,6 +27,24 @@
 /** @typedef {import('./image/parsers/jpeg.js').JpegStartOfFrame} JpegStartOfFrame */
 /** @typedef {import('./image/parsers/jpeg.js').JpegStartOfScan} JpegStartOfScan */
 
+/** @typedef {import('./image/parsers/png.js').PngBackgroundColor} PngBackgroundColor */
+/** @typedef {import('./image/parsers/png.js').PngChromaticities} PngChromaticies */
+/** @typedef {import('./image/parsers/png.js').PngColor} PngColor */
+/** @typedef {import('./image/parsers/png.js').PngCompressedTextualData} PngCompressedTextualData */
+/** @typedef {import('./image/parsers/png.js').PngHistogram} PngHistogram */
+/** @typedef {import('./image/parsers/png.js').PngImageData} PngImageData */
+/** @typedef {import('./image/parsers/png.js').PngImageGamma} PngImageGamma */
+/** @typedef {import('./image/parsers/png.js').PngImageHeader} PngImageHeader */
+/** @typedef {import('./image/parsers/png.js').PngIntlTextualData} PngIntlTextualData */
+/** @typedef {import('./image/parsers/png.js').PngLastModTime} PngLastModTime */
+/** @typedef {import('./image/parsers/png.js').PngPalette} PngPalette */
+/** @typedef {import('./image/parsers/png.js').PngPhysicalPixelDimensions} PngPhysicalPixelDimensions */
+/** @typedef {import('./image/parsers/png.js').PngSignificantBits} PngSignificantBits */
+/** @typedef {import('./image/parsers/png.js').PngSuggestedPalette} PngSuggestedPalette */
+/** @typedef {import('./image/parsers/png.js').PngSuggestedPaletteEntry} PngSuggestedPaletteEntry */
+/** @typedef {import('./image/parsers/png.js').PngTextualData} PngTextualData */
+/** @typedef {import('./image/parsers/png.js').PngTransparency} PngTransparency */
+
 export {
   UnarchiveEvent, UnarchiveEventType, UnarchiveInfoEvent, UnarchiveErrorEvent,
   UnarchiveStartEvent, UnarchiveFinishEvent, UnarchiveProgressEvent, UnarchiveExtractEvent,
@@ -38,6 +56,8 @@ export { GifParseEventType, GifParser } from './image/parsers/gif.js';
 export { JpegComponentType, JpegDctType, JpegDensityUnits, JpegExtensionThumbnailFormat,
          JpegHuffmanTableType, JpegParseEventType, JpegParser,
          JpegSegmentType } from './image/parsers/jpeg.js';
+export { PngColorType, PngInterlaceMethod, PngParseEventType, PngParser,
+  PngUnitSpecifier } from './image/parsers/png.js';
 export { convertWebPtoPNG, convertWebPtoJPG } from './image/webp-shim/webp-shim.js';
 export { BitBuffer } from './io/bitbuffer.js';
 export { BitStream } from './io/bitstream.js';